GithubProfileContract

+++ domain/src/main/java/com/artemissoftware/domain/usecase/GetUserProfileUseCase.kt
@@ -6,27 +6,37 @@ import com.artemissoftware.domain.models.UserProfile
 import com.artemissoftware.domain.repository.GitHubRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
+import java.util.*
 import javax.inject.Inject
 
 class GetUserProfileUseCase @Inject constructor(private val gitHubRepository: GitHubRepository) {
 
     operator fun invoke(name : String): Flow<Resource<UserProfile>> = flow {
 
-        val apiResult = gitHubRepository.getUserProfile(name = name)
+        val userProfile = gitHubRepository.getCachedUserProfile()
 
-        gitHubRepository.deleteCache()
+        if(requestData(userProfile)){
 
-        apiResult.data?.let {
+            val apiResult = gitHubRepository.getUserProfile(name = name)
 
+            apiResult.data?.let {
 
+                gitHubRepository.deleteCache()
+                gitHubRepository.cacheUserProfile(it)
+                emit(Resource.Success(it))
 
-
-            emit(Resource.Success(it))
-
-        } ?: run{
-            throw ApiDataException(apiResult.error)
+            } ?: run{
+                throw ApiDataException(apiResult.error)
+            }
         }
-
+        else{
+            emit(Resource.Success(userProfile!!))
+        }
     }
+
+    private fun requestData(userProfile: UserProfile?): Boolean{
+        return userProfile?.expirationDate?.after(Date())?.not() ?: true
+    }
+
 
 }
